Release v2021.12.06.1 (#1084)

+++ pkg/keyvisual/storage/model_test.go
@@ -74,7 +74,7 @@ func (t *testDbstoreSuite) TestClearTableAxisModel(c *C) {
 	if err != nil {
 		c.Fatalf("Count table AxisModel error: %v", err)
 	}
-	c.Assert(count, Equals, 1)
+	c.Assert(count, Equals, int64(1))
 
 	err = ClearTableAxisModel(t.db)
 	c.Assert(err, IsNil)
@@ -83,7 +83,7 @@ func (t *testDbstoreSuite) TestClearTableAxisModel(c *C) {
 	if err != nil {
 		c.Fatalf("Count table AxisModel error: %v", err)
 	}
-	c.Assert(count, Equals, 0)
+	c.Assert(count, Equals, int64(0))
 }
 
 func (t *testDbstoreSuite) TestAxisModelFunc(c *C) {
@@ -123,7 +123,7 @@ func (t *testDbstoreSuite) TestAxisModelFunc(c *C) {
 	if err != nil {
 		c.Fatalf("Count table AxisModel error: %v", err)
 	}
-	c.Assert(count, Equals, 0)
+	c.Assert(count, Equals, int64(0))
 
 	err = axisModel.Delete(t.db)
 	c.Assert(err, IsNil)
@@ -157,7 +157,7 @@ func (t *testDbstoreSuite) TestAxisModelsFindAndDelete(c *C) {
 	if err != nil {
 		c.Fatalf("Count table AxisModel error: %v", err)
 	}
-	c.Assert(count, Equals, int(maxLayerNum)*axisModelNumEachLayer)
+	c.Assert(count, Equals, int64(int(maxLayerNum)*axisModelNumEachLayer))
 
 	findLayerNum := maxLayerNum - 1
 	axisModels, err := FindAxisModelsOrderByTime(t.db, findLayerNum)
@@ -175,7 +175,7 @@ func (t *testDbstoreSuite) TestAxisModelsFindAndDelete(c *C) {
 	if err != nil {
 		c.Fatalf("Count table AxisModel error: %v", err)
 	}
-	c.Assert(count, Equals, int(maxLayerNum-1)*axisModelNumEachLayer)
+	c.Assert(count, Equals, int64(int(maxLayerNum-1)*axisModelNumEachLayer))
 }
 
 func axisModelsDeepEqual(obtainedAxisModels []*AxisModel, expectedAxisModels []*AxisModel, c *C) {
